requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
Authlib==1.3.0
azure-common==1.1.28
azure-core==1.28.0
azure-functions==1.15.0
azure-identity==1.16.0
azure-search==1.0.0b2
azure-search-documents==11.4.0
azure-storage-blob==12.17.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.6.2
build==0.10.0
CacheControl==0.13.1
cachetools==5.3.3
certifi==2023.11.17
cffi==1.15.0
chardet==5.2.0
charset-normalizer==3.3.2
cleo==2.0.1
click==8.1.5
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.0
coverage==7.2.7
crashtest==0.4.1
cryptography==37.0.2
cycler==0.12.1
dataclasses-json==0.6.3
dataclasses-json-speakeasy==0.5.11
datasets==2.19.1
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
distlib==0.3.7
distro==1.9.0
dnspython==2.4.0
dulwich==0.21.6
effdet==0.4.1
emoji==2.10.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
faiss-cpu==1.7.4
fastapi==0.100.0
filelock==3.12.2
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.47.2
frozenlist==1.4.1
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.64.0
grpcio-health-checking==1.64.0
grpcio-tools==1.64.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.23.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
intel-openmp==2021.4.0
iopath==0.1.10
ipykernel==6.29.0
ipython==8.21.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
keyring==24.2.0
kiwisolver==1.4.5
langchain==0.1.17
langchain-community==0.0.37
langchain-core==0.1.52
langchain-experimental==0.0.58
langchain-openai==0.0.6
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.56
layoutparser==0.3.4
llvmlite==0.41.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl==2021.4.0
more-itertools==10.1.0
mpmath==1.3.0
msal==1.28.0
msal-extensions==1.1.0
msgpack==1.0.7
msrest==0.6.21
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
numba==0.58.1
numexpr==2.8.4
numpy==1.26.3
oauthlib==3.2.0
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.15.1
openai==1.10.0
openai-whisper==20231117
openapi-schema-pydantic==1.2.4
opencv-python==4.9.0.80
openpyxl==3.1.0
orjson==3.10.3
packaging==23.2
pandas==2.2.0
parso==0.8.3
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
pikepdf==8.15.1
pillow==10.3.0
pillow_heif==0.16.0
pkginfo==1.9.6
platformdirs==3.11.0
plotly==5.15.0
pluggy==1.2.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
pony==0.7.17
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==5.27.0
psutil==5.9.8
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pyarrow-hotfix==0.6
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
pydeck==0.9.0
Pygments==2.17.2
PyJWT==2.8.0
pymongo==4.4.1
Pympler==1.0.1
pyparsing==3.1.1
pypdf==4.0.1
PyPDF2==3.0.1
pypdfium2==4.30.0
pyproject_hooks==1.0.0
pyreadline3==3.4.1
pyrsistent==0.19.3
pysbd==0.3.4
pytesseract==0.3.10
pytest==7.4.0
pytest-cov==4.1.0
pytest-cover==3.0.0
pytest-coverage==0.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
pytz==2023.4
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.2
ragas @ git+https://github.com/explodinggradients/ragas.git@e18b927750705a05e7fed7647becc0cadb9aa1bb
rapidfuzz==2.15.2
referencing==0.35.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.18.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
shellingham==1.5.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
stack-data==0.6.3
starlette==0.27.0
store==2020.2.10
streamlit==1.37.1
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tbb==2021.12.0
tenacity==8.2.3
threadpoolctl==3.1.0
tiktoken==0.6.0
timm==1.0.3
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.1
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.30.2
trove-classifiers==2023.9.19
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==4.3.1
unstructured==0.12.4
unstructured-client==0.18.0
unstructured-inference==0.7.23
unstructured.pytesseract==0.3.12
urllib3==2.2.0
uvicorn==0.23.1
validators==0.28.1
virtualenv==20.24.5
watchdog==4.0.0
wcwidth==0.2.13
weaviate-client==4.6.3
Werkzeug==2.3.6
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0